pyproject.toml
[tool.poetry]
name = "baseten-backend-take-home"
version = "1.0.0"
description = "Baseten backend take home python boilerplate"
authors = ["Pierre-Alexandre St-Jean <pa@baseten.co>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.265.0"}
aiohttp = "^3.9.0"
sqlalchemy = "^2.0.0"
aiosqlite = "^0.19.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
